LOCO.API: Create loco_main file and delete old loco
